+++ Tronc commun/Round 2/ft_printf/ft_tools/ft_putchar.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putchar.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: basverdi <basverdi@student.42.fr>          +#+  +:+       +#+        */
+/*   By: bastienverdier-vaissiere <bastienverdie    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/22 15:33:39 by basverdi          #+#    #+#             */
-/*   Updated: 2023/11/22 16:51:05 by basverdi         ###   ########.fr       */
+/*   Updated: 2023/11/24 13:07:35 by bastienverd      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../printf.h"
+#include "../ft_printf.h"
 
-int	ft_putchar(int c)
+size_t	ft_putchar(int c)
 {
 	write(1, &c, 1);
 	return (1);
